Added SPARQL tests for aggregate handling of error/unbound.
Adds two tests for SPARQL handling of aggregation dealing with
error/unbound values:

:agg-count-with-unbound tests COUNT(?o) where some results in the group do not
have a bound value for ?o. This is functionality that was not tested by the
original SPARQL 1.1 test suite (which only tested COUNT over groups where all
values were bound).

:agg-sample-with-unbound tests SAMPLE(?o) where some results in the group do
not have a bound value for ?o. This is functionality which is similarly
untested by the original test suite, and is inspired by the email thread at
[1] which questions how missing variable bindings are handled by SAMPLE.

[1] https://lists.w3.org/Archives/Public/public-sparql-dev/2016JanMar/0005.html